tests/wpt/web-platform-tests/html/editing/focus/inert/inert-does-not-match-disabled-selector.html

<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>inert does not match :disabled selector</title>
    <link rel="author" title="Alice Boxhall" href="aboxhall@chromium.org">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <style>
      button {
        color: green;
      }

      button:disabled {
        color: red;
      }
    </style>
  </head>
  <body>
    <button inert>The test passes if this is in green.</button>
    <script>
      test(function() {
        button = document.querySelector('button');
        var color = document.defaultView.getComputedStyle(button).getPropertyValue('color');
        assert_equals(color, 'rgb(0, 128, 0)');
      }, 'Tests that inert elements do not match the :disabled selector.');
    </script>
  </body>
</html>
